moved url fetchers into article classes
